rename folder Models to lower case  model
