--- Models/tax_config.go
@@ -1,2 +1,0 @@
-package models
-
