+++ services/tax.go
@@ -5,7 +5,7 @@ import (
 	"math"
 
 	"github.com/kanawat2566/assessment-tax/constants"
-	"github.com/kanawat2566/assessment-tax/models"
+	"github.com/kanawat2566/assessment-tax/model"
 	"github.com/kanawat2566/assessment-tax/repository"
 )
